feat(file): ungate the sequential engine core for no_std guests
Add a primitives feature enabling the primitives dependency, the byte buffers and the poll utilities in alloc mode; read, walk, split, store, inflight and num now ride it, with std forwarding to it.

The borrowed-store relay moves to a gated relay module with a cell dual wherever the Send/Sync bounds relax; the multi_thread cfg now clears on bare metal, mirroring the marker traits.

The file and primitives workspace edges drop default features, with consumers enabling std explicitly; the wasm bindings gain the alloc imports the no-std shape needs.

The no_std workflow checks nectar-file with the primitives core on both targets, and the unit workflow runs the pinned-root and split parity batteries without std.

Collapse the rayon cfg trios to the bare feature gate; the compile_error rules already enforce both exclusions.
