+++ Cargo.toml
@@ -62,11 +62,11 @@ check.all_targets = true
 nectar-clock = { version = "0.4.0", path = "crates/clock", default-features = false }
 nectar-contracts = { version = "0.4.0", path = "crates/contracts" }
 nectar-feeds = { version = "0.4.0", path = "crates/feeds" }
-nectar-file = { version = "0.4.0", path = "crates/file" }
+nectar-file = { version = "0.4.0", path = "crates/file", default-features = false }
 nectar-manifest = { version = "0.4.0", path = "crates/manifest" }
 nectar-mantaray = { version = "0.4.0", path = "crates/mantaray" }
 nectar-marker = { version = "0.4.0", path = "crates/marker" }
-nectar-primitives = { version = "0.4.0", path = "crates/primitives" }
+nectar-primitives = { version = "0.4.0", path = "crates/primitives", default-features = false }
 nectar-swarms = { version = "0.4.0", path = "crates/swarms" }
 nectar-tasks = { version = "0.4.0", path = "crates/tasks" }
 nectar-postage = { version = "0.4.0", path = "crates/postage" }
