+++ crates/examples/Cargo.toml
@@ -32,6 +32,6 @@ nectar-contracts.workspace = true
 nectar-postage.workspace = true
 nectar-postage-issuer = { workspace = true, features = ["parallel"] }
 nectar-postage-usage = { workspace = true, features = ["client", "seal"] }
-nectar-primitives.workspace = true
+nectar-primitives = { workspace = true, features = ["std"] }
 thiserror = { workspace = true, features = ["std"] }
 tokio = { workspace = true, features = ["rt", "macros"] }
