+++ crates/file/build.rs
@@ -1,11 +1,13 @@
 //! Emits the `multi_thread` cfg alias.
 
 fn main() {
-    // Send/Sync bounds apply; wasm32 and the `unsync` feature relax them.
+    // Send/Sync bounds apply; wasm32, bare metal and the `unsync` feature
+    // relax them, mirroring the marker traits.
     println!("cargo::rustc-check-cfg=cfg(multi_thread)");
     let wasm32 = std::env::var("CARGO_CFG_TARGET_ARCH").as_deref() == Ok("wasm32");
+    let bare = std::env::var("CARGO_CFG_TARGET_OS").as_deref() == Ok("none");
     let unsync = std::env::var_os("CARGO_FEATURE_UNSYNC").is_some();
-    if !wasm32 && !unsync {
+    if !wasm32 && !bare && !unsync {
         println!("cargo::rustc-cfg=multi_thread");
     }
 }
