+++ crates/manifest-sim/Cargo.toml
@@ -26,7 +26,7 @@ disallowed_methods = "deny"
 [dependencies]
 bytes.workspace = true
 nectar-manifest = { workspace = true }
-nectar-primitives = { workspace = true }
+nectar-primitives = { workspace = true, features = ["std"] }
 nectar-testing.workspace = true
 serde = { workspace = true, features = ["std", "derive"] }
 serde_json = { workspace = true, features = ["std"] }
