+++ crates/primitives/examples/wasm-demo/Cargo.toml
@@ -19,7 +19,7 @@ default = [ "console_error_panic_hook" ]
 
 [dependencies]
 bytes.workspace = true
-nectar-primitives.workspace = true
+nectar-primitives = { workspace = true, features = ["std"] }
 wasm-bindgen.workspace = true
 js-sys.workspace = true
 console_error_panic_hook = { version = "0.1.7", optional = true }
